Add crop button.
I tried cropping the image without scaling the crop_brush dimensions but this doesn't work correctly. I will explain with an example. Suppose the original image is 2000x1000 and the app rescales it to 1000x500 to fit in the window. I select a 200x300 rectangle with the crop guide and click the crop button. Instead of returning a 200x300 cropped area from the scaled 1000x500 image, it returns a 200x300 cropped area from the original 2000x1000 image.

Because we want to crop and save the original-sized image, not the scaled image. To fix the problem, the app calculates the scale factor: scale factor = scaled width / original width. Then the app divides the crop guide dimensions by the scale factor to convert the crop guide dimensions to the original image scale.