Crop rotated image correctly.
magick::image_rotate enlarges the image canvas to fit the rotated image. However, rotating the image does not change the location of the origin (0,0) in the image. Let w1 and h1 be the width and height of the original image respectively and let w2 and h2 be the width and height of the enlarged image canvas. Then the top-left point in the rotated image (enlarged image canvas) is (-(w2-w1)/2, -(h2-h1)/2).

Shiny, on the other hand, moves the origin (0,0) in a rotated image to the top-left corner. In order to crop in the location the user selects in the Shiny app, we need to subtract the (-(w2-w1)/2, -(h2-h1)/2) from in the geometry_area when we use image_crop.