Scaled display image to width of window.
Did. Resized the displayed image to fit the width of the window. The corner points (xmin, xmax, ymin, ymax) of the crop guide are on the same scale as the image resized to fit the width of the window. The point (0,0) is in the top-left corner of the window. We want to crop the oringal-sized, and possibly rotated, image not the resided image, so we multiply the corner points by image$display_width / image$window_width to find the loctions of the corner points on the oringal-sized image.After the corner point locations are found on the original-sized image, the x and y offsets are calculated if the image was rotated.

Did not do. The app returns an error if the user tries to crop a rotated image a second time. I did not fix this.