Only allow an image to be cropped once.
Cropping a cropped and rotated image causes the app to crash. Users shouldn't need to crop a cropped image, so now the app disables the crop button after it is clicked the first time.

The crop button is disabled when the app first loads, and it is enabled if the user selects an area of the image. To keep the crop button from re-enabling after the first crop, I added a logical variable image$display_cropped. If this variable is TRUE, selecting an area of the image will not enable the crop button.